move related word to the navbar
