+++ packages/client/hmi-client/src/components/Navbar.vue
@@ -6,16 +6,18 @@ import Avatar from 'primevue/avatar';
 import Menu from 'primevue/menu';
 import { Project } from '@/types/Project';
 import useAuthStore from '@/stores/auth';
+import Chip from 'primevue/chip';
 import Dialog from 'primevue/dialog';
 import SearchBar from '@/components/data-explorer/search-bar.vue';
 import { RouteMetadata, RouteName } from '@/router/routes';
 import Dropdown from 'primevue/dropdown';
 import { useCurrentRoute, RoutePath } from '@/router/index';
+import { isEmpty } from 'lodash';
+import { getRelatedWords } from '@/services/data';
 
 const props = defineProps<{
 	project: Project | null;
 	searchBarText?: string;
-	relatedSearchTerms?: string[];
 }>();
 
 interface NavItem {
@@ -65,10 +67,6 @@ const userInitials = computed(() =>
 		?.split(' ')
 		.reduce((accumulator, currentValue) => accumulator.concat(currentValue.substring(0, 1)), '')
 );
-
-function searchTextChanged(value) {
-	router.push({ name: RouteName.DataExplorerRoute, query: { q: value } });
-}
 
 function searchByExampleModalToggled() {
 	// TODO
@@ -122,61 +120,107 @@ watch(activeProjectName, (newProjectName) => {
 watch(currentRoute, (newRoute) => {
 	selectedPage.value = navItems.value[newRoute.path] || emptyNavItem;
 });
+
+/*
+ * Search text
+ */
+
+const query = ref<string>();
+
+function searchTextChanged(value) {
+	router.push({ name: RouteName.DataExplorerRoute, query: { q: value } });
+}
+
+/*
+ * Suggested Terms
+ */
+
+const terms = ref<string[]>([]);
+
+watch(
+	() => props.searchBarText,
+	async (newSearchTerm) => {
+		if (newSearchTerm) {
+			terms.value = await getRelatedWords(newSearchTerm);
+		} else {
+			terms.value = [];
+		}
+	},
+	{ immediate: true }
+);
+
+function addSearchTerm(term) {
+	query.value = query.value ? query.value.concat(' ').concat(term) : term;
+}
 </script>
 
 <template>
 	<header>
-		<section class="header-left">
-			<img src="@assets/svg/terarium-logo.svg" height="36" alt="TERArium logo" />
-			<nav>
-				<Dropdown
-					class="dropdown"
-					v-model="selectedPage"
-					:options="Object.values(navItems)"
-					optionLabel="name"
-					panelClass="dropdown-panel"
-					@change="goToPage"
-				>
-					<template #value="slotProps">
-						<i :class="slotProps.value.icon" />
-						<span>{{ slotProps.value.name }}</span>
-					</template>
-					<template #option="slotProps">
-						<i :class="slotProps.option.icon" />
-						<span>{{ slotProps.option.name }}</span>
-					</template>
-				</Dropdown>
-			</nav>
-		</section>
-		<SearchBar
-			class="searchbar"
-			:text="searchBarText"
-			:suggested-terms="relatedSearchTerms"
-			@search-text-changed="searchTextChanged"
-			@toggle-search-by-example="searchByExampleModalToggled"
-		/>
-		<section class="header-right">
-			<Avatar :label="userInitials" class="avatar m-2" shape="circle" @click="showUserMenu" />
-		</section>
-		<Menu ref="userMenu" :model="userMenuItems" :popup="true" />
-		<Dialog header="Logout" v-model:visible="isLogoutDialog">
-			<p>You will be returned to the login screen.</p>
-			<template #footer>
-				<Button label="Cancel" class="p-button-secondary" @click="closeLogoutDialog" />
-				<Button label="Ok" @click="auth.logout" />
-			</template>
-		</Dialog>
+		<main>
+			<section class="header-left">
+				<img src="@assets/svg/terarium-logo.svg" height="36" alt="TERArium logo" />
+				<nav>
+					<Dropdown
+						class="dropdown"
+						v-model="selectedPage"
+						:options="Object.values(navItems)"
+						optionLabel="name"
+						panelClass="dropdown-panel"
+						@change="goToPage"
+					>
+						<template #value="slotProps">
+							<i :class="slotProps.value.icon" />
+							<span>{{ slotProps.value.name }}</span>
+						</template>
+						<template #option="slotProps">
+							<i :class="slotProps.option.icon" />
+							<span>{{ slotProps.option.name }}</span>
+						</template>
+					</Dropdown>
+				</nav>
+			</section>
+			<SearchBar
+				class="searchbar"
+				:text="query"
+				@search-text-changed="searchTextChanged"
+				@toggle-search-by-example="searchByExampleModalToggled"
+			/>
+			<section class="header-right">
+				<Avatar :label="userInitials" class="avatar m-2" shape="circle" @click="showUserMenu" />
+			</section>
+		</main>
+		<aside class="suggested-terms" v-if="!isEmpty(terms)">
+			Suggested terms:
+			<Chip v-for="term in terms" :key="term" removable remove-icon="pi pi-times">
+				<span @click="addSearchTerm(term)">{{ term }}</span>
+			</Chip>
+		</aside>
 	</header>
+	<Menu ref="userMenu" :model="userMenuItems" :popup="true" />
+	<Dialog header="Logout" v-model:visible="isLogoutDialog">
+		<p>You will be returned to the login screen.</p>
+		<template #footer>
+			<Button label="Cancel" class="p-button-secondary" @click="closeLogoutDialog" />
+			<Button label="Ok" @click="auth.logout" />
+		</template>
+	</Dialog>
 </template>
 
 <style scoped>
 header {
-	align-items: center;
 	background-color: var(--surface-section);
-	display: flex;
 	border-bottom: 1px solid var(--surface-border);
 	padding: 0.5rem 1rem;
+}
+
+header > main {
+	align-items: center;
+	display: flex;
 	justify-content: space-between;
+}
+
+header > aside {
+	justify-content: center;
 }
 
 .searchbar {
@@ -215,5 +259,38 @@ header {
 
 i {
 	margin-right: 0.5rem;
+}
+
+/* Suggested terms */
+.suggested-terms {
+	margin: 0.75rem 0 0.25rem 0;
+	display: inline-flex;
+	gap: 0.5rem;
+	margin-left: 0.25rem;
+	margin-right: auto;
+	align-items: center;
+	overflow: hidden;
+	white-space: nowrap;
+}
+
+.suggested-terms,
+.p-chip {
+	font-size: small;
+	font-weight: bold;
+	color: var(--text-color-subdued);
+}
+
+.p-chip {
+	padding: 0 0.75rem;
+	background-color: var(--surface-200);
+}
+
+.p-chip span {
+	margin: 0.25rem 0;
+	cursor: pointer;
+}
+
+.p-chip :deep(.p-chip-remove-icon) {
+	font-size: 0.75rem;
 }
 </style>
